refactor: move Backup testing to test_backup.py
