+++ tests/unit/test_base_backup.py
@@ -21,7 +21,7 @@ from ._testing import _make_credentials
 from google.cloud._helpers import UTC
 
 
-class TestBackup(unittest.TestCase):
+class TestBackupConstants:
     PROJECT_ID = "project-id"
     INSTANCE_ID = "instance-id"
     INSTANCE_NAME = "projects/" + PROJECT_ID + "/instances/" + INSTANCE_ID
@@ -32,11 +32,13 @@ class TestBackup(unittest.TestCase):
     BACKUP_ID = "backup-id"
     BACKUP_NAME = CLUSTER_NAME + "/backups/" + BACKUP_ID
 
+
+class TestBaseBackup(unittest.TestCase, TestBackupConstants):
     @staticmethod
     def _get_target_class():
-        from google.cloud.bigtable.backup import Backup
+        from google.cloud.bigtable.base_backup import BaseBackup
 
-        return Backup
+        return BaseBackup
 
     @staticmethod
     def _make_table_admin_client():
@@ -291,429 +293,9 @@ class TestBackup(unittest.TestCase):
         backup2 = self._make_one("backup_2", "instance2")
         self.assertTrue(backup1 != backup2)
 
-    def test_create_grpc_error(self):
-        from google.api_core.exceptions import GoogleAPICallError
-        from google.api_core.exceptions import Unknown
-        from google.cloud._helpers import _datetime_to_pb_timestamp
-        from google.cloud.bigtable_admin_v2.types import table_pb2
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.create_backup.side_effect = Unknown("testing")
-
-        timestamp = self._make_timestamp()
-        backup = self._make_one(
-            self.BACKUP_ID,
-            _Instance(self.INSTANCE_NAME, client=client),
-            table_id=self.TABLE_ID,
-            expire_time=timestamp,
-        )
-
-        backup_pb = table_pb2.Backup(
-            source_table=self.TABLE_NAME,
-            expire_time=_datetime_to_pb_timestamp(timestamp),
-        )
-
-        with self.assertRaises(GoogleAPICallError):
-            backup.create(self.CLUSTER_ID)
-
-        api.create_backup.assert_called_once_with(
-            parent=self.CLUSTER_NAME, backup_id=self.BACKUP_ID, backup=backup_pb,
-        )
-
-    def test_create_already_exists(self):
-        from google.cloud._helpers import _datetime_to_pb_timestamp
-        from google.cloud.bigtable_admin_v2.types import table_pb2
-        from google.cloud.exceptions import Conflict
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.create_backup.side_effect = Conflict("testing")
-
-        timestamp = self._make_timestamp()
-        backup = self._make_one(
-            self.BACKUP_ID,
-            _Instance(self.INSTANCE_NAME, client=client),
-            table_id=self.TABLE_ID,
-            expire_time=timestamp,
-        )
-
-        backup_pb = table_pb2.Backup(
-            source_table=self.TABLE_NAME,
-            expire_time=_datetime_to_pb_timestamp(timestamp),
-        )
-
-        with self.assertRaises(Conflict):
-            backup.create(self.CLUSTER_ID)
-
-        api.create_backup.assert_called_once_with(
-            parent=self.CLUSTER_NAME, backup_id=self.BACKUP_ID, backup=backup_pb,
-        )
-
-    def test_create_instance_not_found(self):
-        from google.cloud._helpers import _datetime_to_pb_timestamp
-        from google.cloud.bigtable_admin_v2.types import table_pb2
-        from google.cloud.exceptions import NotFound
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.create_backup.side_effect = NotFound("testing")
-
-        timestamp = self._make_timestamp()
-        backup = self._make_one(
-            self.BACKUP_ID,
-            _Instance(self.INSTANCE_NAME, client=client),
-            table_id=self.TABLE_ID,
-            expire_time=timestamp,
-        )
-
-        backup_pb = table_pb2.Backup(
-            source_table=self.TABLE_NAME,
-            expire_time=_datetime_to_pb_timestamp(timestamp),
-        )
-
-        with self.assertRaises(NotFound):
-            backup.create(self.CLUSTER_ID)
-
-        api.create_backup.assert_called_once_with(
-            parent=self.CLUSTER_NAME, backup_id=self.BACKUP_ID, backup=backup_pb,
-        )
-
-    def test_create_cluster_not_set(self):
-        backup = self._make_one(
-            self.BACKUP_ID,
-            _Instance(self.INSTANCE_NAME),
-            table_id=self.TABLE_ID,
-            expire_time=self._make_timestamp(),
-        )
-
-        with self.assertRaises(ValueError):
-            backup.create()
-
-    def test_create_table_not_set(self):
-        backup = self._make_one(
-            self.BACKUP_ID,
-            _Instance(self.INSTANCE_NAME),
-            expire_time=self._make_timestamp(),
-        )
-
-        with self.assertRaises(ValueError):
-            backup.create(self.CLUSTER_ID)
-
-    def test_create_expire_time_not_set(self):
-        backup = self._make_one(
-            self.BACKUP_ID, _Instance(self.INSTANCE_NAME), table_id=self.TABLE_ID,
-        )
-
-        with self.assertRaises(ValueError):
-            backup.create(self.CLUSTER_ID)
-
-    def test_create_success(self):
-        from google.cloud._helpers import _datetime_to_pb_timestamp
-        from google.cloud.bigtable_admin_v2.types import table_pb2
-
-        op_future = object()
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.create_backup.return_value = op_future
-
-        timestamp = self._make_timestamp()
-        backup = self._make_one(
-            self.BACKUP_ID,
-            _Instance(self.INSTANCE_NAME, client=client),
-            table_id=self.TABLE_ID,
-            expire_time=timestamp,
-        )
-
-        backup_pb = table_pb2.Backup(
-            source_table=self.TABLE_NAME,
-            expire_time=_datetime_to_pb_timestamp(timestamp),
-        )
-
-        future = backup.create(self.CLUSTER_ID)
-        self.assertEqual(backup._cluster, self.CLUSTER_ID)
-        self.assertIs(future, op_future)
-
-        api.create_backup.assert_called_once_with(
-            parent=self.CLUSTER_NAME, backup_id=self.BACKUP_ID, backup=backup_pb,
-        )
-
-    def test_exists_grpc_error(self):
-        from google.api_core.exceptions import Unknown
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.get_backup.side_effect = Unknown("testing")
-
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-
-        with self.assertRaises(Unknown):
-            backup.exists()
-
-        api.get_backup.assert_called_once_with(self.BACKUP_NAME)
-
-    def test_exists_not_found(self):
-        from google.api_core.exceptions import NotFound
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.get_backup.side_effect = NotFound("testing")
-
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-
-        self.assertFalse(backup.exists())
-
-        api.get_backup.assert_called_once_with(self.BACKUP_NAME)
-
-    def test_get(self):
-        from google.cloud.bigtable_admin_v2.gapic import enums
-        from google.cloud.bigtable_admin_v2.proto import table_pb2
-        from google.cloud._helpers import _datetime_to_pb_timestamp
-
-        timestamp = _datetime_to_pb_timestamp(self._make_timestamp())
-        state = enums.Backup.State.READY
-
-        client = _Client()
-        backup_pb = table_pb2.Backup(
-            name=self.BACKUP_NAME,
-            source_table=self.TABLE_NAME,
-            expire_time=timestamp,
-            start_time=timestamp,
-            end_time=timestamp,
-            size_bytes=0,
-            state=state,
-        )
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.get_backup.return_value = backup_pb
-
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-
-        self.assertEqual(backup.get(), backup_pb)
-
-    def test_reload(self):
-        from google.cloud.bigtable_admin_v2.gapic import enums
-        from google.cloud.bigtable_admin_v2.proto import table_pb2
-        from google.cloud._helpers import _datetime_to_pb_timestamp
-
-        timestamp = _datetime_to_pb_timestamp(self._make_timestamp())
-        state = enums.Backup.State.READY
-
-        client = _Client()
-        backup_pb = table_pb2.Backup(
-            name=self.BACKUP_NAME,
-            source_table=self.TABLE_NAME,
-            expire_time=timestamp,
-            start_time=timestamp,
-            end_time=timestamp,
-            size_bytes=0,
-            state=state,
-        )
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.get_backup.return_value = backup_pb
-
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-
-        backup.reload()
-        self.assertEqual(backup._source_table, self.TABLE_NAME)
-        self.assertEqual(backup._expire_time, timestamp)
-        self.assertEqual(backup._start_time, timestamp)
-        self.assertEqual(backup._end_time, timestamp)
-        self.assertEqual(backup._size_bytes, 0)
-        self.assertEqual(backup._state, state)
-
-    def test_exists_success(self):
-        from google.cloud.bigtable_admin_v2.proto import table_pb2
-
-        client = _Client()
-        backup_pb = table_pb2.Backup(name=self.BACKUP_NAME)
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.get_backup.return_value = backup_pb
-
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-
-        self.assertTrue(backup.exists())
-
-        api.get_backup.assert_called_once_with(self.BACKUP_NAME)
-
-    def test_delete_grpc_error(self):
-        from google.api_core.exceptions import Unknown
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.delete_backup.side_effect = Unknown("testing")
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-
-        with self.assertRaises(Unknown):
-            backup.delete()
-
-        api.delete_backup.assert_called_once_with(self.BACKUP_NAME)
-
-    def test_delete_not_found(self):
-        from google.api_core.exceptions import NotFound
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.delete_backup.side_effect = NotFound("testing")
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-
-        with self.assertRaises(NotFound):
-            backup.delete()
-
-        api.delete_backup.assert_called_once_with(self.BACKUP_NAME)
-
-    def test_delete_success(self):
-        from google.protobuf.empty_pb2 import Empty
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.delete_backup.return_value = Empty()
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-
-        backup.delete()
-
-        api.delete_backup.assert_called_once_with(self.BACKUP_NAME)
-
-    def test_update_expire_time_grpc_error(self):
-        from google.api_core.exceptions import Unknown
-        from google.cloud._helpers import _datetime_to_pb_timestamp
-        from google.cloud.bigtable_admin_v2.types import table_pb2
-        from google.protobuf import field_mask_pb2
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.update_backup.side_effect = Unknown("testing")
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-        expire_time = self._make_timestamp()
-
-        with self.assertRaises(Unknown):
-            backup.update_expire_time(expire_time)
-
-        backup_update = table_pb2.Backup(
-            name=self.BACKUP_NAME, expire_time=_datetime_to_pb_timestamp(expire_time),
-        )
-        update_mask = field_mask_pb2.FieldMask(paths=["expire_time"])
-        api.update_backup.assert_called_once_with(
-            backup_update, update_mask,
-        )
-
-    def test_update_expire_time_not_found(self):
-        from google.api_core.exceptions import NotFound
-        from google.cloud._helpers import _datetime_to_pb_timestamp
-        from google.cloud.bigtable_admin_v2.types import table_pb2
-        from google.protobuf import field_mask_pb2
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.update_backup.side_effect = NotFound("testing")
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-        expire_time = self._make_timestamp()
-
-        with self.assertRaises(NotFound):
-            backup.update_expire_time(expire_time)
-
-        backup_update = table_pb2.Backup(
-            name=self.BACKUP_NAME, expire_time=_datetime_to_pb_timestamp(expire_time),
-        )
-        update_mask = field_mask_pb2.FieldMask(paths=["expire_time"])
-        api.update_backup.assert_called_once_with(
-            backup_update, update_mask,
-        )
-
-    def test_update_expire_time_success(self):
-        from google.cloud._helpers import _datetime_to_pb_timestamp
-        from google.cloud.bigtable_admin_v2.proto import table_pb2
-        from google.protobuf import field_mask_pb2
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.update_backup.return_type = table_pb2.Backup(name=self.BACKUP_NAME)
-        instance = _Instance(self.INSTANCE_NAME, client=client)
-        backup = self._make_one(self.BACKUP_ID, instance, cluster_id=self.CLUSTER_ID)
-        expire_time = self._make_timestamp()
-
-        backup.update_expire_time(expire_time)
-
-        backup_update = table_pb2.Backup(
-            name=self.BACKUP_NAME, expire_time=_datetime_to_pb_timestamp(expire_time),
-        )
-        update_mask = field_mask_pb2.FieldMask(paths=["expire_time"])
-        api.update_backup.assert_called_once_with(
-            backup_update, update_mask,
-        )
-
-    def test_restore_grpc_error(self):
-        from google.api_core.exceptions import GoogleAPICallError
-        from google.api_core.exceptions import Unknown
-
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.restore_table.side_effect = Unknown("testing")
-
-        timestamp = self._make_timestamp()
-        backup = self._make_one(
-            self.BACKUP_ID,
-            _Instance(self.INSTANCE_NAME, client=client),
-            cluster_id=self.CLUSTER_ID,
-            table_id=self.TABLE_NAME,
-            expire_time=timestamp,
-        )
-
-        with self.assertRaises(GoogleAPICallError):
-            backup.restore(self.TABLE_ID)
-
-        api.restore_table.assert_called_once_with(
-            parent=self.INSTANCE_NAME, table_id=self.TABLE_ID, backup=self.BACKUP_NAME,
-        )
-
-    def test_restore_cluster_not_set(self):
-        client = _Client()
-        client.table_admin_client = self._make_table_admin_client()
-        backup = self._make_one(
-            self.BACKUP_ID,
-            _Instance(self.INSTANCE_NAME, client=client),
-            table_id=self.TABLE_ID,
-            expire_time=self._make_timestamp(),
-        )
-
-        with self.assertRaises(ValueError):
-            backup.restore(self.TABLE_ID)
-
-    def test_restore_success(self):
-        op_future = object()
-        client = _Client()
-        api = client.table_admin_client = self._make_table_admin_client()
-        api.restore_table.return_value = op_future
-
-        timestamp = self._make_timestamp()
-        backup = self._make_one(
-            self.BACKUP_ID,
-            _Instance(self.INSTANCE_NAME, client=client),
-            cluster_id=self.CLUSTER_ID,
-            table_id=self.TABLE_NAME,
-            expire_time=timestamp,
-        )
-
-        future = backup.restore(self.TABLE_ID)
-        self.assertEqual(backup._cluster, self.CLUSTER_ID)
-        self.assertIs(future, op_future)
-
-        api.restore_table.assert_called_once_with(
-            parent=self.INSTANCE_NAME, table_id=self.TABLE_ID, backup=self.BACKUP_NAME,
-        )
-
 
 class _Client(object):
-    def __init__(self, project=TestBackup.PROJECT_ID):
+    def __init__(self, project=TestBackupConstants.PROJECT_ID):
         self.project = project
         self.project_name = "projects/" + self.project
